+++ src/main/java/org/sspoid/sspoid/api/dto/auth/KakaoUserInfoResponse.java
@@ -1,10 +1,9 @@
-package org.sspoid.sspoid.api.dto;
+package org.sspoid.sspoid.api.dto.auth;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import org.springframework.context.annotation.Profile;
 
 import java.util.HashMap;
