+++ src/main/java/org/sspoid/sspoid/api/dto/auth/KakoTokenResponse.java
@@ -1,4 +1,4 @@
-package org.sspoid.sspoid.api.dto;
+package org.sspoid.sspoid.api.dto.auth;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
